RPG/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
flake8==7.0.0
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==8.2.0
python-dotenv==1.0.1
requests==2.32.2
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
WTForms==3.1.2
